requirements.txt
Flask==2.2.5
pymongo==4.11.2
python-dotenv==1.0.1
Werkzeug==3.1.3
